Reset 3G modem if present but no connection for some time
This should help with cases where a connection has never been
established because the 3G modem is locked up (as opposed to the 3G
interface being up but not being functional). Action is only taken if
the newer PCB is in use (and the power can just be pulled from the 3G
modem) to avoid continual reboots in areas where there's no 3G
coverage.